feat: Establish test extension and shims library (#8)
Establishes an extension that entrains SES with eventual send (`E()`)
for the purpose of tests, experimentation, and prototyping. This is
accomplished by the addition of two new monorepo packages:
- `@ocap/extension`
- `@ocap/shims`

These packages are private. The `@ocap` prefix is merely a convenience
serving to identify monorepo packages from external packages. We are not
going to claim this prefix in any public registry, and any packages that
we want to publish in the future will be renamed `@metamask`.

Various changes are made to config files in the root workspace in
service of the new packages. In particular, `constraints.pro` has been
modified, and a number of exceptions have been added. However, reviewers
are encouraged to ignore this file, since Prolog constraints are
deprecated in Yarn 4, and we will eventually migrate to its JavaScript
constraints system.

Added code follows existing MetaMask conventions to the greatest extent
possible. Both packages are exceptional in that neither of them are
TypeScript libraries, and the extension of course isn't a library at
all. Therefore, neither package necessarily sets a precedent for a
"conventional" package in this monorepo.

## `@ocap/shims`

This package contains various shims that we want to import as-is,
without applying any transforms. This includes `ses` and `lockdown`.
Moreover, some of these shims, such as `eventual-send`, need to be
applied in a particular order to be successfully applied. To this end,
the package exports `endoify.mjs`, which is intended to be The One Shim
to Rule Them All<sup>TM</sup> for ocap programming purposes.

Note that, while the eventual send shim is applied, eventual send itself
(`E()`) isn't actually used anywhere yet. We will definitively use it in
the immediate future, however, so we may as well add it now.

## `@ocap/extension`

This package consists of a headless
[MV3](https://developer.chrome.com/docs/extensions/develop/migrate/what-is-mv3)
Chrome extension. See its readme for installation and usage
instructions.

The extension has three entry-points:

- `background.js`, for its background service worker
- `offscreen.html`, for its [offscreen
document](https://developer.chrome.com/docs/extensions/reference/api/offscreen)
- `iframe.html`, for the iframes that we append to the offscreen
document
- These iframes are [sandboxed
pages](https://developer.chrome.com/docs/extensions/how-to/security/sandboxing-eval#load_file),
for whose CSP we are allowed to enable `unsafe-eval`, and thereby
`Compartment.evaluate()` (from the SES shim).
- This is the [recommended
way](https://developer.chrome.com/docs/extensions/reference/api/offscreen)
to access `unsafe-eval` in MV3 extensions.
- This is also what we do for Snaps in the production extension, see
https://github.com/MetaMask/metamask-extension/pull/25171.

The extension is written in TypeScript where possible, and bundled using
[Vite](https://vitejs.dev/). Shims and other files are statically
imported (i.e. without bundling) using `vite-plugin-static-copy`. Easily
testable modules are tested using [Vitest](https://vitest.dev/), Vite's
Jest-inspired and _ESM-compatible_ test runner.