chore: Reset `yarn.lock` (#26)
Does a hard reset of our lockfile by doing `rm yarn.lock && yarn
install`. This is to solve [a recurring CI
failure](https://github.com/MetaMask/ocap-kernel/actions/runs/10452768542/job/28941881443?pr=16)
of unknown provenance. This nuclear option was the only thing I tried
that was able to resolve the problem. If we were in production we would
be more hesitant to make such a move, but in our current state it
doesn't really matter.