feat(shims): Bundle endoify dependencies (#34)
A note on testing the shims package in the ocap monorepo:

The CI build job requires that yarn build succeeds with node 20.
The CI test job (essentially) requires that yarn test succeeds with
node 18 and 20. The test job does not have access to the build
artifacts from the build job, which allows the build and test jobs
to run in parallel.

In order to allow the shims package to test its built shims,
we prefix the shims test command with `yarn build && `, which, as
a command called during the CI test job, runs with both node 18
and 20. Therefore, although our CI suggests the intended workflow
for the ocap monorepo is "build with 20, run with 18 or 20", the
tests tighten the constraint in the shims package specifically to
"build and run with 18 or 20".