build(extension): Robustify JavaScript module externalization (#45)
Closes: #14 

We require that our trusted computing base be endoified. Our strategy
for meeting this requirement is largely encapsulated in
`@ocap/shims/endoify.js`, but also includes maintaining the following
strict import procedure for all javascript entry points.

1. Import trusted code
2. Import endoify
3. Import untrusted code

We call steps 1 and 2 in order the _trusted prelude_ of a file.

Even if we properly set up all the source code for our entry points
with a trusted prelude, it is still possible for the bundling process to
interfere by permuting the order of import statements in the output
files. Such a permutation effectively moves untrusted code into the
trusted computing base, which our plans cannot tolerate.

To address this possibility, we break out entry point foo's trusted prelude
into a file called foo-trusted-prelude which is imported in foo's source
and explicitly declared the trusted prelude of foo in the configuration of
the jsTrustedPrelude vite plugin.

The jsTrustedPrelude plugin is a vite plugin to ensure the following.
- Every declared trusted prelude is handled externally.
- Every declared trusted prelude importer:
  - Is a declared entry point,
  - Imports at most one declared trusted prelude,
  - Begins by importing its declared trusted prelude.