build(extension): Enable source maps (#59)
This enables source maps for the extension, which are disabled in Vite
by default.

Vite has a whole concept of "modes", so that you can do e.g. `vite
--mode production` and `vite --mode development`. By default everything
runs in "production", and so do we. However, this is not a meaningful
distinction for us at the moment or possibly ever, so we can just add
whatever options we find pleasant. That of course includes source maps.