chore: Update lint configuration (#65)
- Fixes Prettier `--ignore-path` in package `lint:misc` scripts
- Reconfigures root package lint scripts to run `eslint` and `prettier`
directly from the root
- Both of these applications support monorepos, and there's no need to
invoke them in parallel from each package. Moreover, if we do not do it
this way, the root directory will not be linted.
- I have never observed any differences from running `prettier` or
`eslint` from the root vs. within a package, but should we ever find
that they in some way conflict, we can revert to the parallel invocation
with a single invocation for the root directory, excluding `/packages`.
- Ensures that dependency linting is run for the monorepo root, and adds
a `lint:dependencies:fix` script that runs `depcheck` and dedupes
dependencies.