build(extension): Add build-time type checks (#77)
Adds build-time type checks to Vite via `vite-plugin-checker`, which we
configure to run `tsc` as part of the Vite pipeline.

As part of this, the `lint:types` script introduced in #73 (f.k.a.
`build:types`, which was a misnomer) is removed. In essence, I first
attempted to establish a `lint:types` script common across all packages.
Any monorepo-wide implementation thereof turns out to be equivalent to
just running a full, declaration-only TypeScript build of the entire
monorepo (see [the TypeScript project references
documentation](https://www.typescriptlang.org/docs/handbook/project-references.html)
for details). This being the case, rather than using a one-off script
for `@ocap/extension`, I found a Vite plugin that allows us to check
types as part of the build process, as we do everywhere else.