feat(streams,extension): Add `chrome.runtime` streams (#128)
Adds streams of `chrome.runtime.onMessage` / `.sendMessage` and
integrates them into the extension. This transport mechanism uses what
appears to be `JSON.stringify()` to serialize its messages, and
therefore does not support sending native `Error` objects. Consequently,
we also introduce a custom marshaling layer to the base stream class, in
order to forward errors that are passed to `.throw()`. We forego the use
of `@endo/marshal` for now, since it doesn't support e.g. `error.cause`.
The marshaling functions can always be parameterized to the stream
classes in the future.

The error marshaling required a lot of runtime validation, for which we
entrain
[`@metamask/superstruct`](https://github.com/MetaMask/superstruct). This
is used to generate TypeScript type and create type guards elsewhere in
the organization, and it will probably grow to replace our existing type
guards.

Finally, minimizes the unit tests for both stream implementations.
Stream implementations should only test their usage of the transport
mechanism, and the handoff to the base stream. The in-depth testing of
stream semantics we should keep in `BaseStream.test.ts`.