chore: Fix type import specifier linting (#175)
This PR reconfigures our `tsconfig` files and ESLint config to enforce
consistent type import specifiers and better type elision behavior.

Heretofore, we have enforced the existence of type import specifiers,
but that they are stylistically consistent. The options are "top-level"
(`import type { a } from 'x'}`) and "inline" (`import { type a } from
'x'}`). This was enforced using the ESLint rule
`@typescript-eslint/consistent-type-imports`. This rule has an option
named `fixTypes`, but that only applies when a type-only import is _not_
specified as a type-only import. In other words, it enforces the
existence of type import specifiers, but not that they're stylistically
consistent.

After consulting the [rule
documentation](https://typescript-eslint.io/rules/consistent-type-imports/#comparison-with-importsnotusedasvalues--verbatimmodulesyntax),
the [TypeScript
docs](https://www.typescriptlang.org/tsconfig/#verbatimModuleSyntax),
and https://github.com/typescript-eslint/typescript-eslint/issues/9652,
I determined that we are better off enabling `verbatimModuleSyntax` in
our `tsconfig` files, and replacing
`@typescript-eslint/consistent-type-imports` with
`import-x/consistent-type-specifier-style`. In this way, the existence
of type import specifiers is enforced by `tsc`, and the `import-x`
ESLint plugins ensures that they are consistent. We use the "top-level"
style because that results in `import type { a } from 'x';` being elided
completely from the JS output as opposed to `import { type a } from
'x';` which would be emitted as `import {} from 'x';`. I don't know if
that's actually harmful, but it seems surprising, and I don't like
surprises.