refactor(streams)!: Stop sending iterator results over the wire (#187)
Refactors the internals of `BaseReader` and `BaseWriter` such that they
no longer send or expect `IteratorResult` values over the wire.
Previously, we enshrined the `MarshaledError` error type and the `{
done: true, value: undefined }` iterator result as the "throw" and
"done" signals for the stream. Now, we have established two special
values of the type `StreamSignal` to handle these cases. They are
completely internal to the streams, and they are not observable using
the public stream API.

Removing the expectation that `Dispatchable` values must conform to
`MarshaledError` or `IteratorResult` results in less wrapping of values
on the wire. Meanwhile, the introduction of dedicated stream signals
disambiguates "normal" traffic from internal control signals, and
enables us to change their types independently.