feat(streams): Synchronize duplex streams  (#189)
Introduces a notion of "synchronization" to duplex streams, such that
they must receive an acknowledgement from their remote counterpart
before they can begin processing reads and writes. This is similar to
the `cork`/`uncork` dyad in Node.js streams. The protocol recapitulates
the one used for the same purpose in `@metamask/post-message-stream`,
which is itself modeled on TCP's SYN / SYN-ACK / ACK handshake.

This new functionality is integrated into the kernel and extension.
Regardless of when and in what order two paired duplex streams are
constructed, message loss should be impossible. This also enables
simplifications of existing logic added to work around the possibility
of message loss.