chore: Move lint:ts check from pre-commit to GitHub workflow (#195)
Since `yarn lint:ts` depends on the build process, running it during
pre-commit would force developers to build before committing, which is
not ideal for efficiency. This PR removes the `lint:ts` check from the
`pre-commit` hook and adds it to the GitHub workflow, where it will be
executed as part of the build step. It was moved on build step because
caching the build folder for workflows took [2min on the
ci](https://github.com/MetaMask/ocap-kernel/actions/runs/11504219003/job/32023298354)
so it was faster to move it after build.

---------

Co-authored-by: Erik Marks <25517051+rekmarks@users.noreply.github.com>