Implement kernel storage abstraction layer (#180)
This is a first cut at a storage wrapper for the kernel. What I have
assumes it is passed a dumb string/string key/value store, and builds on
top of that, sometimes in ways that are necessarily a bit indirect.
Obviously, if we have a more featureful storage substrate we could
potentially take advantage of its features to do this stuff more
directly (such a thing would also imply refactoring the construction API
to push the responsibility for creating whatever object or objects
actually accesses the storage to a lower level in the code, rather than
simply passing in the kv store at the top).

It's also worth mentioning that at this point there is not yet any
notion of transaction boundaries, which will absolutely be needed but
which is more wrapped up in the specifics of the underlying storage
layer than the stuff here is.