feat(streams): Add consumer-provided input validation to streams (#219)
Ref #218 

Adds consumer-provided input validation to streams in the form of a
function constructor argument. This is necessary in order for streams to
actually enforce the types they claim to read and write. Input
validation only occurs in read streams, because we can simply rely on
TypeScript to write the correct values. If input validation fails, the
read stream will end.