feat(streams): Add stream multiplexer (#221)
Ref: #218 

Adds a new class `StreamMultiplexer` to `@ocap/streams`. A multiplexer
is not a stream itself, but rather a wrapper around a duplex stream. The
multiplexer provides methods for creating "channels" over the underlying
stream, which are themselves duplex streams and may have a different
message type and validation logic.

The multiplexer is constructed in an idle state, and must be explicitly
"started" via the `start()` or `drainAll()` methods. All channels must
be added before the multiplexer is started.

Starting the multiplexer will synchronize the underlying duplex stream,
if it is synchronizable. Therefore, in order to prevent message loss,
callers **should not** synchronize the underlying duplex stream before
passing it to the multiplexer. For the same reason, the multiplexer will
throw if any channels are added after it has started. We could enable
"dynamic" channel creation by synchronization the channels themselves,
but only if there's a clear need.

See tests for example usage.