refactor: Use streams for vat worker service communication (#246)
Replaces ad hoc messaging implementation with streams for the
`ExtensionVatWorkerClient` and `*Server`. This is accomplished by
refactoring the `PostMessage` stream classes to handle transferables
over the underlying `postMessage` transport. Implementing both this
capability and multiplexing simultaneously would require a significant
amount of extra work, and multiplexing is instead removed for this
stream class for now. (The `PostMessage` stream had in fact fallen out
of use prior to this PR.)

As part of these changes, improves test coverage for the vat worker
service classes, and removes the integration tests established
therefore. Finally, makes some small changes to `test-utils` in service
of a better experience when writing tests.

The outcome of all of this is that we are now using streams for all IPC
in the extension. Time will tell whether this effort has been worth it.