chore: Launch kernel panel on start command (#277)
Closes #69 

This PR adds a new Vite plugin that automatically manages the browser
instance during `yarn start`. When the development server starts, the
plugin launches Chrome with the extension loaded and opens the extension
popup. As code changes are made, it automatically reloads the extension
and reopens the popup window, eliminating the need to manually reload
the extension or reopen the popup during development. The plugin also
ensures clean shutdown by closing the browser instance when the dev
server stops. Lastly, the extension details page is launched before the
popup, so one can have easier access to devtools for the extension.