refactor: Move `delay()` to production utils (#282)
We originally added a `delay()` function to `test-utils`, but we have
equivalent functionality in production code. Therefore, this function
now lives in the propduction `utils` package. We continue to inline a
declaration of this function in the `extension-dev` Vite plugin because
the `utils` package expects a hardened environment.

Also, deletes `makeHandledCallback()`, which had become unused.