chore: Add `create-package` utility (#295)
Adds a `create-package` utility via a script of the same name, modeled
on https://github.com/MetaMask/core/pull/3581. The implementation is a
copypaste of
https://github.com/MetaMask/core/tree/85188d8146d13f06e1fe1a4143b6a35ac95630ea/scripts/create-package,
adapted for our needs. This should henceforth be used to add new
packages to the monorepo.

We have one manual step specific to us, namely adding coverage
thresholds for the new package to the root `vitest.config.ts`. The
disadvantage of TypeScript config files is that they're hard to work
with programmatically. We experimented with breaking the thresholds out
into a separate JSON file, but then threshold auto-updating stopped
working. Ah, well, nevertheless.

To test, run something like the following command, and observe that a
new package `@ocap/foobar` is created:

```text
y create-package -n foobar -d "A foobar package."
```