chore: Run Dependabot for all packages on a weekly schedule (#298)
Until now, we have used the same schedule / configuration for Dependabot
as other MetaMask repositories. Namely, we only bump packages from a set
of trusted publishers, including ourselves, on a daily basis. This is
because bumping all packages can be distracting, as well as a security
risk due to supply chain attacks. Since we are pre-production, we can
afford to live closer to the bleeding edge, and we will therefore try
bumping all our packages on a weekly basis.