feat: Integrate kernel with liveslots (#317)
This PR implements the first phase of integration between the
`@agoric/swingset-liveslots` package and our kernel.

This enables full vat-to-vat messaging, including serialization and
deserialization of all passable value types, intervat object reference
mapping, and cross-vat promise resolution. This means we now have a
minimum viable multi-vat computational platform!

The word "minimal" there is important: there remain a few rough edges
and a number of important things that still need to be done, which
should be converted into issues of their own in our project plan.

Notable points:

- This PR adds kernel support for the vat syscall and liveslots delivery
interfaces, but while the skeleton for these is complete, the kernel
does not yet implement the elements of these interfaces that realize
distributed reference GC. Without this, the memory footprint of the vats
and the kernel will grow over time without bound, which is, of course,
bad, but it's entirely adequate for things like demos, proofs of
concept, and of course as the jumping off point for actually
implementing the DGC stuff. Issue #329 has been added for this.

- The liveslots vat syscall interface is nominally a synchronous API,
whereas the kernel<->vat message pathway is unavoidably asynchronous on
account of how web APIs work. However, the only portions of this API
that truly require synchronous invocation are those that pertain to vat
storage access. We deal with this by implementing the vat storage
portions of the syscall API directly in the vat process as part of the
`VatSupervisor` class. Basically, each vat gets its own SQLite database
for its state. One consequence of this is that, for the time being at
least, vats run on top of a memory-only version of SQLite, which means
the nominally persistent vat state is not actually persistent; our
discussions have convinced us that for a lot of applications this
ephemerality is actually acceptable. However, some applications will
require genuine vat persistence. The way to achieve this is to run them
inside web workers instead of iframes, where SQLite can make use of the
synchronous part of the OPFS API (as the kernel does). However, the
version of the code in this PR does not include support for vats in web
workers; that will need to be added. Issue #332 has been added for this.

- An additional point about wrapping the synchronous syscall interface
around the asynchronous vat<->kernel interface: one reason we can get
away with this is that, aside from the vat storage methods already
discussed, the results returned by syscall methods largely don't matter
-- the only thing returned is an error indication in the case of
something going wrong, and none of these calls can go wrong as the
result of user error. In other words, they should all always be
successful except in the case of bugs in our own code. So the
`VatSupervisor` implements these by shipping off the asynchronous
syscall request to the kernel and then synchronously returning a success
indicator to liveslots. This really should be OK, but probably some
better logging is called for in order to catch bugs in our stuff in the
event we have any (like that could ever happen, hah). Issue #333 has
been added for this.

- In the same vein, in the interest of minimizing unanticipated
implementation shear due to expectations built into liveslots, this PR
implements the very serial one-crank-at-a-time model that SwingSet
currently does. However, I don't think this is strictly necessary. In
particular, since vats each have their own storage, they can implement
internal message queues and then be allowed to run freely, with all of
the side effects of a delivery being released to the kernel in a single
blob upon crank completion. This is an opportunity for better
parallelism and likely better performance, as well as giving us more
protection from what happens if a piece of user code in a vat goes into
an infinite loop. However, to support all this will require significant
additions to both the vat supervisor and the kernel, so I'm kicking that
can down the road for now. For the record I've added issue #334 for
this.

- Although the current vat store implementation is ephemeral, it should
still support the complete virtual object and virtual collection APIs
(e.g., exoclasses, baggage, et al) that liveslots makes available to vat
user code. All that stuff should Just Work, modulo the fact that the
bits will evaporate with their enclosing iframe. However, while I have
high confidence that all this will, in fact, Just Work, none of this
functionality has actually yet been tested in our environment, and you
know what they say about untested code. Issue #335 has been added for
this.

- All of the new intervat computation stuff and its supporting machinery
implemented here currently lacks unit tests. This is an important gap,
but I'm putting this PR up now anyway in the interest of making it all
available to the whole team and of avoiding too much drift from the main
branch. Issue #336 as been added for this.

- There were a number of issues with the `@agoric` packages as published
not playing well with TypeScript. These issues are on the Agoric side,
having to do with how they build and publish this stuff to npm. The
Agoric folks are aware and supportive, but we will probably have to be
more proactive on our side with respect to getting that stuff fixed.
Everybody's expectation is that this is just some tedious build and
configuration stuff that needs to be sorted out, but it's possible that
we will end up having to be the ones to do it via a PR to Agoric. In the
meantime, I've worked around these issues via the expedient of
incorporating the necessary type declaration files into our own source
tree in the `kernel` package. These were created by extracting the
relevant declaration files from the Agoric sources in the `agoric-sdk`
repo and then having Claude convert them from JSDoc to actual TypeScript
(all these files have names beginning with `ag-` if you want to look at
them). Once the npm packaging issues are dealt with, these can go away.
Having our own type declaration files is bad mainly due to the risk of
divergence if there are changes on the Agoric side; I'm not too worried
about this as these APIs have been stable for some time, but still...
I've added issue #337 for this.

- Quite a lot of code has been added in a fairly short time, so there
are certainly some code hygiene issues that should be addressed. I'm
usually pretty obsessive about that kind of stuff, but, as with unit
tests, in the interest of moving our process along I'm pushing now
rather than taking additional time to make it all perfect and pretty
first. In particular, I think there are some methods in `Kernel`,
`VatHandle`, and `VatSupervisor` that should be private but aren't.
Also, some additional love ought to be applied to the layout of the
source files in terms of how things are ordered for the benefit of human
readers. Also, more comments.

- While we're on the subject of code hygiene, this exercise has really
driven home to me just how clunky, awkward, ugly, and needlessly verbose
doing things in the JS class pattern is, in contrast to the
objects-as-closures pattern. I think I'd like to raise this as an
internal issue in our team at some point.

- I'd give even odds that there are one or two other notable points that
I should be mentioning here but which I've lost track of in all the
excitement. When and if I think of them I'll add additional comments to
this PR.

Closes #29