chore: Clean up and document `dummyMeterControl` (#348)
This cleans up and documents `dummyMeterControl` per #341. In the course
of doing this, a deeper look at how liveslots uses the meter control
object revealed that some of the mechanism in the original Agoric
version of this, which I just adopted wholesale and blindly translated
to TypeScript, are actually required despite the fact that my initial
assessment was that they were not and could be removed/simplified.
However, the main thrust of the original complaint that prompted issue
#341 was lack of clarity as to this file's function, so the issue is
still relevant. I've cleaned up a couple of minor code infelicities and
documented everything.

Closes #341