Eliminate some vestigial remnants of kernel development (#365)
This PR removes the `kvGet`, `kvSet`, and `evaluate` messages from the
kernel and vat message protocols. These were basically test scaffolding
that were helpful for debugging during early development but which have
outlived their usefulness. These operations shouldn't be exposed on a
public interface anyway. I left in the `ping` message for now, as I
suspect it may be useful for some sanity checks and is probably not so
sensitive.