Simplify `Ref` types (#366)
This simplifies the `Ref` type hierarchy (`Ref`, `ERef`, `VRef`, `RRef`,
and `KRef`; also `VatId` and `EndpointId` while I was at it). The
original types incorporated some of the syntax rules for the strings
that characterize these references, but while that was clever in theory,
it proved awkward and cumbersome in practice, necessitating an annoying
number of type casts to make `tsc` happy while at the same time
providing little actual benefit in terms of correctness. This PR reduces
the leaf types of the type hierarchy to simple strings, and eliminates
quite a large number of type casts.

Interestingly, this uncovered a couple of cases where the casts,
intended to address conversion from string to ref type, were masking the
need for runtime checks for `null` or `undefined`. These checks have
been inserted. Note that this PR does _not_ address some places where we
properly ought to be doing runtime validity checks of the actual
reference strings themselves, but we already have an issue (#338) for
these.

Closes #345