refactor: Replace kernel-UI communication with `BroadcastChannel` (#349)
Closes #280 

Establishes a `BroadcastChannel` channel between the UI and kernel
worker for kernel-UI messages. This removes the necessity of routing
this traffic through the offscreen page, and entirely removes the need
for stream multiplexing, which is removed. In order to handle reloading
the UI (which closes and then reopens the stream and `BroadcastChannel`
on that end), extends the duplex stream synchronization protocol to
support "re-synchronization". In consequence, the UI no longer returns
its stream before it closes, in order to keep the remote (i.e. kernel)
side open for when the UI reopens.

In addition:
- ensures that `DuplexStream.drain()` is completely orthogonal to
synchronization (i.e. it can be called at any time, before, during, or
after (re-)synchronization).
- uses `jsdom` and the mock endo environment for `@ocap/streams` unit
tests.
  - The devex isn't worth it.

This currently prevents us from having multiple UI instances at once, but
we will fix this in #371.