fix(kernel): Remove misplaced endoify import (#372)
We were importing the "endoify" shim in `Kernel.ts`, but it should only
be imported in entry point files in platform runtimes (i.e. `nodejs` and
`extension`). Although the import is not harmful at runtime, importing
it in non-entry point files makes it difficult to mock during testing
and development.