Increased Kernel parameter check paranoia (#389)
This adds extra runtime parameter checking in the kernel.

As I was doing this, to make sense of some stuff that was going on I
also ended up cleaning up some naming confusion. to wit: the name
`sendMessage` was confusingly overloaded, with different meanings in
different contexts. I changed one of these contexts to use the name
`sendVatCommand`, which clarified a bunch of things but ended up
touching a surprising number of files. I considered making that a
separate PR but in the end decided the extra effort of doing that didn't
pay for itself.

Closes issue #338