chore: Bump TypeScript to 5.8 (#425)
Bumps TypeScript to version 5.8. This required a minor refactor of
TypeScript configuration in order to continue resolving source files at
dev time.

The version bumped to is an RC release, and our `typescript-eslint`
doesn't yet support this version of TypeScript. Normally, we would wait
for the entire ecosystem to catch up, but due to work stacked on top of
this PR that will go stale, we'll live with the minor risk and bump the
relevant packages again later.