Test package for running cross-vat tests.  And tests too. (#418)
This adds a new `kernel-test` package that has the machinery for running
tests that involve message traffic being exchanged between vats.

Also included are a bunch of tests for various cases of liveslots <->
kernel promise handling.

Note that this probably needs some build script love. While it has a
bunch of vat code for running the various tests, the build process does
not generate the bundles that the tests themselves expect to see. Right
now this needs to be done manually, but really should be automated.

Closes #336