refactor: Enable erasure-only TypeScript (#427)
Requires #426 

Enables the `erasureOnlySyntax` compiler option for all TypeScript
configs. This ensures that compiling our TypeScript to JavaScript never
transforms our source code, and instead just "strips" all type
information. This allows our TypeScript files to be directly executed by
the Node runtime, and allows us to get rid of `tsx`. In addition, it
will improve our interoperability with Agoric, who are planning to begin
using erasure-only TypeScript.

Note that we modify the file extensions of local imports in
`create-package` from `.js` to `.ts` because Node expects the import
specifiers to match those that are actually on disk. This is documented
in the relevant `tsconfig.json`.