chore: Use actual source file extensions for relative imports (#430)
In #427 we started using the actual source file extension for relative
imports in `create-package`. This is because we never actually build
that package, and execute it directly with `node`. After some research,
it turns out that we can do this everywhere using TS 5's
`allowImportingTsExtensions` along with the more recently introduced
`rewriteRelativeImportExtensions`. This finally lets us use the actual,
complete on-disk source file names for imports, and let TypeScript /
Vite worry about what they should be after compilation.