refactor: Standardize construction of objects with async side effects (#432)
Closes #138 

Our codebase has multiple objects that need to perform some set of
asynchronous operations upon or shortly after construction. Originally,
we solved this problem in an ad hoc fashion wherever it appeared. Here
we attempt to standardize our approach to these creational async side
effects.

In general, creational async side effects that (1) need to occur before
their parents can be used and (2) _can_ occur immediately after their
parents are instantiated should go in a static async factory method
`make()`. This is now the case for the `Kernel` and `VatHandle` classes,
which previously had a public `init()` method that had to be called
after construction.

The other case is the `ExtensionVatWorkerClient` and `*Server`, which
have a separate `start()` method. This method returns a promise that
only settles if the operation of these objects fail, and we want to
reserve it for error handling purposes. I attempted to encapsulate this
into `make()` in #433, but decided it was not worth the cost of adding a
compound return value to that method. In the words of @FUDCo, sometimes
the side effects _want_ to be separate from initialization. This is one
of those cases. Since we only use these objects internally and in one
location each, we will simply read our own docstrings and not misuse
them.