refactor: Remove extraneous uses of TypeScript's DOM lib (#435)
We were out of a reliance on defaults injecting TypeScript's DOM lib
into a bunch of packages that don't actually need it. Using this lib in
packages that will run in non-browser environments is dangerous, because
you may come to rely on platform features that won't exist at runtime.
Accomplishing this ~was mostly trivial~ turned out to be painful, but
doable. Our source code is mostly unchanged, but our tooling required
two significant changes:

- Add a `/browser` entrypoint to `@ocap/streams`.
- Without this refactor, TypeScript complains about missing browser
types in packages that omit the DOM lib and import `@ocap/streams`.
- Remove the `lint:ts` scripts and `tsconfig.lint.json` files.
- For whatever reason, this script became dependent on fresh builds,
which is unacceptable for linting. We can live without it.