fix(store): use ephemeral ':memory:' for default store (#446)
If you want things persisted, you must explicitly state where.

This PR coordinates the wasm and nodejs store implementations to use the
same default store location, and sets that default location to
':memory:', which indicates an ephemeral storage location that is lost
on process termination.

Explicitly states the storage location for various tests and for the
extension kernel worker.